+++ src/main/java/nl/mpcjanssen/simpletask/Simpletask.java
@@ -96,6 +96,9 @@ public class Simpletask extends ListActivity  {
 	private final static int DRAWER_CONTEXT = 1;
 	private final static int DRAWER_PROJECT = 2;
 
+
+
+
 	private TaskBag taskBag;
 
 	Menu options_menu;
@@ -563,6 +566,12 @@ public class Simpletask extends ListActivity  {
 		for (Task t : tasks) {
 			if (t != null && !t.isCompleted()) {
 				t.markComplete(new Date());
+                // Recurring task?
+                if (t.getRecurrencePattern()!=null) {
+                    Task newTask = new Task(0,t.getOriginalText());
+                    newTask.deferDueDate(t.getRecurrencePattern());
+                    taskBag.addAsTask(newTask.inFileFormat());
+                }
 			}
 		}
 		if (m_app.isAutoArchive()) {
@@ -590,21 +599,23 @@ public class Simpletask extends ListActivity  {
 		sendBroadcast(new Intent(getPackageName()+Constants.BROADCAST_START_SYNC_TO_REMOTE));
 	}
 
-	private void deferTasks(List<Task> tasks) {
+	private void deferTasks(List<Task> tasks, final int dateType ) {
 		String[] keys = getResources().getStringArray(R.array.deferOptions);
-		SimpleDateFormat formatter = new SimpleDateFormat(
-				Constants.DATE_FORMAT, Locale.US);
-		Date now = new Date();
 		final List<Task> tasksToDefer = tasks;
-		String today = formatter.format(now);
-		String tomorrow = formatter.format(Util.addWeeksToDate(now, 1));
-		String oneWeek = formatter.format(Util.addWeeksToDate(now, 7));
-		String twoWeeks = formatter.format(Util.addWeeksToDate(now, 14));
-		String oneMonth = formatter.format(Util.addMonthsToDate(now, 1));
+		String today = "0d";
+		String tomorrow = "1d";
+		String oneWeek = "1w";
+		String twoWeeks = "2w";
+		String oneMonth = "1m";
 		String[] values = { today, tomorrow, oneWeek, twoWeeks, oneMonth, "" };
-
+        int titleId;
+        if (dateType==Task.DUE_DATE) {
+            titleId = R.string.defer_due;
+        } else {
+            titleId = R.string.defer_threshold;
+        }
 		Dialog d = Util.createSingleChoiceDialog(this, keys, values, 2,
-				R.string.defer, null, new Util.OnSingleChoiceDialogListener() {
+				titleId, null, new Util.OnSingleChoiceDialogListener() {
 					@Override
 					public void onClick(String selected) {
 						if (selected.equals("")) {
@@ -613,7 +624,7 @@ public class Simpletask extends ListActivity  {
 								public void onDateSet(DatePicker datePicker, int year, int month, int day) {
 									Calendar cal = Calendar.getInstance();
 									cal.set(year, month, day);
-									deferTasks(cal.getTime(), tasksToDefer);
+									deferTasks(cal.getTime(), tasksToDefer, dateType);
 
 								}
 							},
@@ -623,17 +634,21 @@ public class Simpletask extends ListActivity  {
 
 							dialog.show();
 						} else {
-							deferTasks(selected, tasksToDefer);
+							deferTasks(selected, tasksToDefer, dateType);
 						}
 					}
 				});
 		d.show();
 	}
 
-	private void deferTasks(Date selected, List<Task> tasksToDefer) {
+	private void deferTasks(Date selected, List<Task> tasksToDefer, int type) {
 		for (Task t : tasksToDefer) {
 			if (t != null) {
-				t.deferToDate(m_app.isDeferThreshold(), selected);
+                if (type==Task.DUE_DATE) {
+				    t.setDueDate(selected);
+                } else {
+                    t.setThresholdDate(selected);
+                }
 			}
 		}
 		m_adapter.setFilteredTasks(false);
@@ -645,10 +660,14 @@ public class Simpletask extends ListActivity  {
 	}
 
 
-	private void deferTasks(String selected, List<Task> tasksToDefer) {
+	private void deferTasks(String selected, List<Task> tasksToDefer, int type) {
 		for (Task t : tasksToDefer) {
 			if (t != null) {
-				t.deferToDate(m_app.isDeferThreshold(), selected);
+                if (type==Task.DUE_DATE) {
+				    t.deferDueDate(selected);
+                } else {
+                    t.deferThresholdDate(selected);
+                }
 			}
 		}
 		m_adapter.setFilteredTasks(false);
@@ -1549,9 +1568,12 @@ public class Simpletask extends ListActivity  {
 				case R.id.delete:
 					deleteTasks(checkedTasks);
 					break;
-				case R.id.defer:
-					deferTasks(checkedTasks);
-					break;
+				case R.id.defer_due:
+                    deferTasks(checkedTasks, Task.DUE_DATE);
+                    break;
+                case R.id.defer_threshold:
+                    deferTasks(checkedTasks, Task.THRESHOLD_DATE);
+                    break;
 				case R.id.uncomplete:
 					undoCompleteTasks(checkedTasks);
 					break;